41.13 Users can delete only their own ideas

+++ app/Http/Livewire/DeleteIdea.php
@@ -4,6 +4,7 @@ namespace App\Http\Livewire;
 
 use App\Models\Idea;
 use App\Models\Vote;
+use Illuminate\Http\Response;
 use Livewire\Component;
 
 class DeleteIdea extends Component {
@@ -14,6 +15,11 @@ class DeleteIdea extends Component {
     }
 
     public function deleteIdea() {
+        #Authorization
+        if(auth()->guest() || auth()->user()->cannot('delete', $this->idea)){
+            abort(Response::HTTP_FORBIDDEN);
+        }
+
         Vote::where('idea_id', $this->idea->id)->delete();
         Idea::destroy($this->idea->id);
